Extract elementwise kernel to a header
